Moved roll to a dice folder to match core folder layout.
